requirements.txt
#Augmentor
Cython
flask
flask-BasicAuth
flask-HTTPAuth
imgaug
ipywidgets
ipykernel
jupyter
keras
matplotlib
opencv-python
pillow
tensorflow==2.3.1
pandas
scikit-learn